move main process file
