+++ packages/desktop/lib/electron/menus/menu-state.constant.ts
@@ -1,7 +1,7 @@
 export const MENU_STATE = {
     enabled: true,
     strings: {
-        about: 'About Firefly',
+        about: 'About Bloom',
         checkForUpdates: 'Check for Updates',
         settings: 'Settings',
         general: 'General',
