+++ packages/desktop/lib/electron/menus/wallet.menu.ts
@@ -1,4 +1,4 @@
-import { getOrInitWindow } from '../main'
+import { getOrInitWindow } from '../processes/main.process'
 import { MENU_STATE } from './menu-state.constant'
 
 export const walletMenu: Electron.MenuItemConstructorOptions = {
